feat(main.go,actions): add command to generate a changelog entry for an individual repo (#47)
* feat(main.go,actions): add command to generate a changelog entry for an individual repo

This PR also adds two new packages for changelog generation and git SHA manipulation, respectively. It does not, however, refactor the existing global changelog generation code to use the new packages in an effort to keep the entire scope of the change down. The work to refactor the global changelog code is forthcoming and represented in https://github.com/deis/deisrel/issues/51